Jooq meter map markers

See merge request elvaco/mvp!1121

+++ backend/configuration/src/integrationTest/java/com/elvaco/mvp/web/MapMarkerControllerTest.java
@@ -24,6 +24,7 @@ import com.elvaco.mvp.web.dto.MapMarkerWithStatusDto;
 import com.elvaco.mvp.web.dto.MapMarkersDto;
 import com.google.common.collect.ImmutableMultimap;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -37,6 +38,7 @@ import static com.elvaco.mvp.core.spi.data.RequestParameter.CITY;
 import static com.elvaco.mvp.testing.fixture.LocationTestData.kungsbacka;
 import static java.util.UUID.randomUUID;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.Assume.assumeTrue;
 
 public class MapMarkerControllerTest extends IntegrationTest {
 
@@ -51,11 +53,18 @@ public class MapMarkerControllerTest extends IntegrationTest {
   @Autowired
   private MeterAlarmLogs meterAlarmLogs;
 
+  @Before
+  public void setUp() {
+    assumeTrue(isPostgresDialect());
+  }
+
   @After
   public void tearDown() {
-    meterAlarmLogJpaRepository.deleteAll();
-    gatewayStatusLogJpaRepository.deleteAll();
-    gatewayJpaRepository.deleteAll();
+    if (isPostgresDialect()) {
+      meterAlarmLogJpaRepository.deleteAll();
+      gatewayStatusLogJpaRepository.deleteAll();
+      gatewayJpaRepository.deleteAll();
+    }
   }
 
   @Test
